Move generic code into helpers
