+++ hometown/web/src/Hometown.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { render, unmountComponentAtNode } from 'react-dom'
+import { wrapComponent } from './wrapComponent'
 
 interface HomeTownPageProps {
   country: string
@@ -15,49 +15,10 @@ const HometownPage: React.FC<HomeTownPageProps> = ({
   return <>{`${city}, ${stateProv}, ${country}`}</>
 }
 
-export class Hometown extends HTMLElement {
-  private observer: MutationObserver
-  private mountPoint: HTMLElement | null = null
-
-  constructor() {
-    super()
-
-    this.observer = new MutationObserver(() => this.update())
-    this.observer.observe(this, { attributes: true })
-  }
-
-  connectedCallback() {
-    this.mount()
-  }
-
-  disconnectedCallback() {
-    this.unmount()
-    this.observer.disconnect()
-  }
-
-  mount() {
-    const root = this.shadowRoot ?? this.attachShadow({ mode: 'open' })
-    this.mountPoint = document.createElement('div')
-    root.appendChild(this.mountPoint)
-
-    render(
-      <HometownPage
-        country={this.getAttribute('country') ?? ''}
-        stateProv={this.getAttribute('stateProv') ?? ''}
-        city={this.getAttribute('city') ?? ''}
-      />,
-      this.mountPoint
-    )
-  }
-
-  unmount() {
-    unmountComponentAtNode(this.mountPoint ?? this)
-  }
-
-  update() {
-    this.unmount()
-    this.mount()
-  }
-}
-
-customElements.define('hometown-page', Hometown)
+wrapComponent('hometown-page', (element) => (
+  <HometownPage
+    country={element.getAttribute('country') ?? ''}
+    stateProv={element.getAttribute('stateProv') ?? ''}
+    city={element.getAttribute('city') ?? ''}
+  />
+))
